==> R/01_data_import.R <==
### Libraries, helper functions and data import

## Load libraries

library(tidyverse)
library(sf)
library(ggspatial)
library(extrafont)


## Import data

load("data/WH.Rdata")
load("data/WH_streets.Rdata")
load("data/WH_territories.Rdata")
load("data/WH_property.Rdata")
load("data/WH_all_property.Rdata")
load("data/WH_territories_property.Rdata")
load("data/WH_ML_property.Rdata")
load("data/WH_LTM_property.Rdata")
load("data/WH_daily.Rdata")
load("data/WH_territories_daily.Rdata")
load("data/WH_ML_daily.Rdata")
load("data/WH_FREH.Rdata")
load("data/WH_GH.Rdata")
